@ physigym : in the middle of next generation implementation.

+++ test/test_physigym_model.py
@@ -185,13 +185,13 @@ class TestPhysigymTib(object):
     def test_physigym_tib_classic(self):
         # install model
         os.chdir(s_path_physigym)
-        o_result = subprocess.run(["python3", "install_physigym.py", "tumor_immune_base", "-f"], check=False, capture_output=True)
+        o_result = subprocess.run(["python3", "install_physigym.py", "tibbue", "-f"], check=False, capture_output=True)
         os.chdir(s_path_physicell)
         # rest output folder
         shutil.rmtree("output/", ignore_errors=True)
         # load, compile, and run model
         o_result = subprocess.run(['make', 'data-cleanup', 'clean', 'reset'], check=False, capture_output=True)
-        o_result = subprocess.run(['make', 'load', 'PROJ=physigym_tumor_immune_base'], check=False, capture_output=True)
+        o_result = subprocess.run(['make', 'load', 'PROJ=physigym_tibbue'], check=False, capture_output=True)
         o_result = subprocess.run(['sed', '-ie ', r's/<max_time units="min">[0-9.]*<\/max_time>/<max_time units="min">1440.0<\/max_time>/g', 'config/PhysiCell_settings.xml'], check=False, capture_output=True)
         o_result = subprocess.run(['sed', '-ie ', r's/<omp_num_threads>[0-9]*<\/omp_num_threads>/<omp_num_threads>4<\/omp_num_threads>/g', 'config/PhysiCell_settings.xml'], check=False, capture_output=True)
         o_result = subprocess.run(['sed', '-ie ', r's/<random_seed>.*<\/random_seed>/<random_seed>system_clock<\/random_seed>/g', 'config/PhysiCell_settings.xml'], check=False, capture_output=True)
@@ -211,16 +211,16 @@ class TestPhysigymTib(object):
         """ note: be hooked up to the internet to run this test successfully. """
         # install model
         os.chdir(s_path_physigym)
-        o_result = subprocess.run(["python3", "install_physigym.py", "tumor_immune_base", "-f"], check=False, capture_output=True)
+        o_result = subprocess.run(["python3", "install_physigym.py", "tibbue", "-f"], check=False, capture_output=True)
         os.chdir(s_path_physicell)
         # rest output folder
         shutil.rmtree("output/", ignore_errors=True)
         # load, compile, and run model
         o_result = subprocess.run(['make', 'data-cleanup', 'clean', 'reset'], check=False, capture_output=True)
-        o_result = subprocess.run(['make', 'load', 'PROJ=physigym_tumor_immune_base'], check=False, capture_output=True)
+        o_result = subprocess.run(['make', 'load', 'PROJ=physigym_tibbue'], check=False, capture_output=True)
         o_result = subprocess.run(['make'], check=False, capture_output=True)
-        o_result = subprocess.run(['python3','custom_modules/physigym/physigym/envs/run_physigym_tib_episodes.py', '--max_time', '1440.0', '--thread', '4', '--seed', 'None'], check=False, capture_output=True)
-        #print("\n", o_result)
+        o_result = subprocess.run(['python3','custom_modules/physigym/physigym/envs/run_physigym_tibbue_episodes.py', '--max_time', '1440.0', '--thread', '4', '--seed', 'None'], check=False, capture_output=True)
+        print("\n", o_result)
         # test for output
         assert(os.path.exists("output/episode00000000/output00000023.xml")) and \
               (os.path.exists("output/episode00000001/output00000023.xml")) and \
@@ -234,33 +234,33 @@ class TestPhysigymTib(object):
         s_unittest = "test_physigym_tib_rl"
         # install model
         os.chdir(s_path_physigym)
-        o_result = subprocess.run(["python3", "install_physigym.py", "tumor_immune_base", "-f"], check=False, capture_output=True)
+        o_result = subprocess.run(["python3", "install_physigym.py", "tibbue", "-f"], check=False, capture_output=True)
         os.chdir(s_path_physicell)
         # rest output folder
         shutil.rmtree("output/", ignore_errors=True)
-        for s_dir in os.listdir("tensorboard/"):
-            if (s_dir.startswith(s_unittest)):
-                shutil.rmtree(f"tensorboard/{s_dir}")
+        if os.path.exists("tensorboard/"):
+            for s_dir in os.listdir("tensorboard/"):
+                if (s_dir.startswith(s_unittest)):
+                    shutil.rmtree(f"tensorboard/{s_dir}")
         # load, compile, and run model
         o_result = subprocess.run(['make', 'data-cleanup', 'clean', 'reset'], check=False, capture_output=True)
-        o_result = subprocess.run(['make', 'load', 'PROJ=physigym_tumor_immune_base'], check=False, capture_output=True)
-        o_result = subprocess.run(['sed', '-ie ', r's/<max_time units="min">[0-9.]*<\/max_time>/<max_time units="min">1440.0<\/max_time>/g', 'config/PhysiCell_settings.xml'], check=False, capture_output=True)
-        o_result = subprocess.run(['sed', '-ie ', r's/<omp_num_threads>[0-9]*<\/omp_num_threads>/<omp_num_threads>4<\/omp_num_threads>/g', 'config/PhysiCell_settings.xml'], check=False, capture_output=True)
-        o_result = subprocess.run(['sed', '-ie ', r's/<random_seed>.*<\/random_seed>/<random_seed>system_clock<\/random_seed>/g', 'config/PhysiCell_settings.xml'], check=False, capture_output=True)
-        o_result = subprocess.run(['sed', '-ie ', rf's/name: str = ".*"/name: str = "{s_unittest}"/g', 'custom_modules/physigym/physigym/envs/sac_tib.py'], check=False, capture_output=True)
-        o_result = subprocess.run(['sed', '-ie ', r's/cuda:.*bool.*=.*True/cuda: bool = False/g', 'custom_modules/physigym/physigym/envs/sac_tib.py'], check=False, capture_output=True)
-        o_result = subprocess.run(['sed', '-ie ', r's/wandb_track:.*bool.*=.*True/wandb_track: bool = False/g', 'custom_modules/physigym/physigym/envs/sac_tib.py'], check=False, capture_output=True)
-        o_result = subprocess.run(['sed', '-ie ', r's/total_timesteps:.*int.*=.*int(.*)/total_timesteps: int = int(72)/g', 'custom_modules/physigym/physigym/envs/sac_tib.py'], check=False, capture_output=True)
-
+        o_result = subprocess.run(['make', 'load', 'PROJ=physigym_tibbue'], check=False, capture_output=True)
+        # bue
+        #o_result = subprocess.run(['python3','custom_modules/physigym/physigym/envs/run_physigym_tibbue_sac.py', '--max_time_episode', '1440.0', '--thread', '4', '--seed', 'None', '--observation_mode', 'img_mc', '--render_mode', 'None', '--name', s_unittest, '--wandb', 'False', '--total_step_learn', '72'], check=False, capture_output=True)
+        #o_result = subprocess.run(['python3','custom_modules/physigym/physigym/envs/run_physigym_tibbue_sac.py', '--max_time_episode', '1440.0', '--thread', '4', '--seed', 'None', '--observation_mode', 'img_rgb', '--render_mode', 'rgb_array', '--name', s_unittest, '--wandb', 'False', '--total_step_learn', '72'], check=False, capture_output=True)
+        o_result = subprocess.run(['python3','custom_modules/physigym/physigym/envs/run_physigym_tibbue_sac.py', '--max_time_episode', '1440.0', '--thread', '4', '--seed', 'None', '--observation_mode', 'scalars', '--render_mode', 'None', '--name', s_unittest, '--wandb', 'False', '--total_step_learn', '72'], check=False, capture_output=True)
+        # alex
+        #o_result = subprocess.run(['sed', '-ie ', rf's/name: str = ".*"/name: str = "{s_unittest}"/g', 'custom_modules/physigym/physigym/envs/sac_tib.py'], check=False, capture_output=True)
+        #o_result = subprocess.run(['sed', '-ie ', r's/wandb_track:.*bool.*=.*True/wandb_track: bool = False/g', 'custom_modules/physigym/physigym/envs/sac_tib.py'], check=False, capture_output=True)
+        #o_result = subprocess.run(['sed', '-ie ', r's/total_timesteps:.*int.*=.*int(.*)/total_timesteps: int = int(72)/g', 'custom_modules/physigym/physigym/envs/sac_tib.py'], check=False, capture_output=True)
         o_result = subprocess.run(['make'], check=False, capture_output=True)
-        o_result = subprocess.run(['python3','custom_modules/physigym/physigym/envs/sac_tib.py'], check=False, capture_output=True)
-        #print("\n", o_result)
+        print("\n", o_result)
         # test for output
-        assert(os.path.exists("output/output00000023.xml")) and \
-              (os.path.exists("tensorboard/")) and \
+        #assert(os.path.exists("output/output00000023.xml")) and \
+        assert(os.path.exists("tensorboard/")) and \
               (any([s_dir.startswith(s_unittest) for s_dir in os.listdir("tensorboard/")]))
         # reset output folder
-        shutil.rmtree("output/")
-        for s_dir in os.listdir("tensorboard/"):
-            if (s_dir.startswith(s_unittest)):
-                shutil.rmtree(f"tensorboard/{s_dir}")
+        #shutil.rmtree("output/")
+        #for s_dir in os.listdir("tensorboard/"):
+        #    if (s_dir.startswith(s_unittest)):
+        #        shutil.rmtree(f"tensorboard/{s_dir}")
